dijkstra was realized for : OrderTypeLayers(Linear,Parallel),BranchCodeLayers(Linear,Parallel)(with computePathes method),for all List of orders(Linear). Usability was checked with tests.
