+++ results/BKLinear.txt
@@ -1254,3 +1254,59 @@
  Path :: [[-37.8158343,145.04645] , [-37.81214,145.01541] , [-37.8115789,145.0113915] , [-37.8103197,144.9983089] , [-37.8102664,144.9977363] , [-37.8097457,144.9913426] , [-37.80892,144.98255] , [-37.8081091,144.9749603] , [-37.8079398,144.9733751] , [-37.80792,144.97319] , [-37.8079064,144.9730355] , [-37.80773,144.97136] , [-37.8050873,144.959954] , [-37.8044544,144.9596573] , [-37.8042161,144.9595444] , [-37.8041766,144.959177] , [-37.804139,144.9588278] , [-37.8041243,144.9586853] , [-37.8041053,144.9585577] , [-37.8040967,144.9585081] , [-37.804091,144.9584652] , [-37.8040854,144.9584151] , [-37.8040695,144.9582857] , [-37.8040517,144.9581406] , [-37.8039764,144.9574981] , [-37.803965,144.9574007] , [-37.8038546,144.9564585] , [-37.8038433,144.9563621] , [-37.8038392,144.9563271] , [-37.8038253,144.9562081] , [-37.8037609,144.9556014] , [-37.8037441,144.9554409] , [-37.8036897,144.9549765] , [-37.8036314,144.9544783] , [-37.8035927,144.954148] , [-37.8036161,144.9541139] , [-37.8036134,144.9539699] , [-37.803578,144.9539284] , [-37.8035245,144.953434] , [-37.8034374,144.9526287] , [-37.803422,144.9524861] , [-37.8033534,144.9518514] , [-37.8032848,144.9512167] , [-37.8033206,144.9511536] , [-37.8033053,144.9510182] , [-37.8032633,144.9509739] , [-37.8032271,144.9506548] , [-37.8031877,144.9503077] , [-37.8031607,144.9500698] , [-37.8031147,144.9496683] , [-37.8031013,144.9495505] , [-37.8031001,144.949536] , [-37.8030948,144.9494928] , [-37.8030545,144.9491339] , [-37.8030255,144.9488558] , [-37.8029931,144.9485699] , [-37.8029575,144.9482791] , [-37.8029482,144.9481976] , [-37.8033457,144.9481274] , [-37.803466,144.9481039] , [-37.803499,144.948358] , [-37.8031708,144.9484204] ]
  Distance: Calculated -> 8975.0 In Dataset -> 8975.0 Epsilon -> 0.0
 
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
