+++ results/NSLinear.txt
@@ -1374,3 +1374,63 @@
  Path :: [[-37.7738026,144.9836466] , [-37.8051496,144.961486] , [-37.8051813,144.9614811] , [-37.8113458,144.9669715] , [-37.8006096,144.9702257] , [-37.800561,144.9701393] , [-37.8007919,144.9701024] , [-37.8008493,144.9700932] , [-37.8009521,144.9700611] , [-37.8012086,144.970017] , [-37.8021231,144.9698599] , [-37.8030239,144.9697046] , [-37.8030832,144.9696944] , [-37.8033526,144.9696479] , [-37.8040317,144.9695353] , [-37.8045175,144.9694499] , [-37.8049263,144.969378] , [-37.8052332,144.9693261] , [-37.8071864,144.9689935] , [-37.8072639,144.968907] , [-37.8073495,144.9688889] , [-37.8079972,144.9687948] , [-37.8080663,144.9688152] , [-37.8081719,144.9688582] , [-37.8082148,144.9687138] , [-37.8082347,144.9686435] , [-37.8082488,144.9685936] , [-37.8084634,144.9679881] , [-37.8085022,144.9680061] , [-37.8088683,144.966779] , [-37.8089463,144.9665017] , [-37.8089655,144.9664333] , [-37.8090083,144.9662934] , [-37.8090317,144.9662126] , [-37.8095148,144.9645449] , [-37.8095522,144.9645624] , [-37.8099666,144.9645969] , [-37.809982,144.9645405] , [-37.8104749,144.9643799] , [-37.8105488,144.9643375] , [-37.8114301,144.9647417] , [-37.811443,144.9647049] , [-37.8114481,144.9646868] , [-37.811453,144.9646701] , [-37.8114581,144.9646549] , [-37.8114868,144.9645604] , [-37.8115745,144.9642719] , [-37.8113564,144.9641723] , [-37.811156,144.9640807] , [-37.8109425,144.9639832] , [-37.8106855,144.9638658] ]
  Distance: Calculated -> 7749.0 In Dataset -> 7536.0 Epsilon -> 213.0
 
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
