+++ results/TPLinear.txt
@@ -1334,3 +1334,51 @@
  Path :: [[-37.8618349,144.905716] , [-37.819966,144.9654451] , [-37.8201239,144.9646328] , [-37.8203215,144.9640179] , [-37.8203326,144.9640349] , [-37.820351,144.9640389] , [-37.8204423,144.9640738] , [-37.8205952,144.9635094] , [-37.8207161,144.9630537] , [-37.8208834,144.9624074] , [-37.8209026,144.9623358] , [-37.8210938,144.9618111] , [-37.8212111,144.9615682] , [-37.8212159,144.9615429] , [-37.8212301,144.9614747] , [-37.8212372,144.961431] , [-37.821239,144.9614195] , [-37.8212464,144.9613796] , [-37.8209341,144.9613851] , [-37.8209342,144.9613737] , [-37.8209345,144.9613277] , [-37.8200187,144.9612976] , [-37.8199672,144.9612505] , [-37.8200685,144.9607348] , [-37.8198669,144.9601208] , [-37.819903,144.9600059] , [-37.8197751,144.9599386] , [-37.8197185,144.9599089] , [-37.8196477,144.9599071] , [-37.819616,144.9599063] , [-37.8195447,144.9599034] , [-37.8195118,144.9598887] , [-37.8191206,144.9597134] , [-37.8186777,144.959515] , [-37.8177544,144.9591013] , [-37.8176866,144.9590709] , [-37.817659,144.9590581] , [-37.8176318,144.9590456] , [-37.8176006,144.9590312] , [-37.8175633,144.9590139] , [-37.8171932,144.9588418] , [-37.8166544,144.9585949] , [-37.8162933,144.9584283] , [-37.8158055,144.9582047] , [-37.8157263,144.9581669] , [-37.8156873,144.9581489] , [-37.8156592,144.9581359] , [-37.8156236,144.9581195] , [-37.8155781,144.9580985] , [-37.814691,144.9576895] , [-37.8138504,144.9573018] , [-37.8137717,144.9572655] , [-37.8136757,144.9572213] , [-37.8136191,144.9571952] , [-37.8127376,144.9567887] , [-37.8118993,144.9564023] , [-37.8118053,144.9563588] , [-37.811764,144.9563397] , [-37.8117362,144.9563269] , [-37.8117249,144.9563218] , [-37.8117008,144.9563112] , [-37.8116841,144.956303] , [-37.811661,144.9562924] , [-37.8106893,144.9558442] , [-37.8103821,144.9557026] , [-37.810106,144.9555752] , [-37.8097306,144.9553906] , [-37.8096098,144.9552809] , [-37.8086347,144.9548232] , [-37.808604,144.954759] , [-37.808531,144.9547279] , [-37.8084394,144.9547952] , [-37.8083317,144.9547551] , [-37.8078888,144.9545472] , [-37.8077923,144.9545028] , [-37.8076815,144.9544519] , [-37.8070061,144.9541677] , [-37.806898,144.9541113] , [-37.8062072,144.9538029] , [-37.8056673,144.9536238] , [-37.8055185,144.95227] , [-37.8055026,144.9521252] , [-37.8053499,144.9508028] , [-37.805337,144.950642] , [-37.8052516,144.9498659] , [-37.8054831,144.9498234] , [-37.8056133,144.9499433] ]
  Distance: Calculated -> 9796.0 In Dataset -> 9772.0 Epsilon -> 24.0
 
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
+ [0.0,0.0] 
+ Path :: []
+ Distance: Calculated -> 1.7976931348623157E308 In Dataset -> NaN Epsilon -> NaN
+
